Added monitor.lua, which makes cybermon perform in the way it used to before this whole LUA thing i.e. it reports when big quantities of data are sent to an attacker.

+++ ip.C
@@ -265,7 +265,7 @@ void ip::process_ip4(engine& eng, context_ptr c, pdu_iter s, pdu_iter e)
 
     else {
 	std::ostringstream buf;
-	buf << "IP protocol " << protocol << " not handled.";
+	buf << "IP protocol " << (int) protocol << " not handled.";
 	throw exception(buf.str());
     }
